Moving base classes to /lib/classes/base
