+++ lib/categories/CoreCategory.js
@@ -1,4 +1,4 @@
-const BaseCategory = require("./BaseCategory.js");
+const BaseCategory = require("../classes/base/BaseCategory.js");
 
 class CoreCategory extends BaseCategory {
 	static code = "basic";
