+++ lib/categories/GuildCategory.js
@@ -1,4 +1,4 @@
-const BaseCategory = require("./BaseCategory.js");
+const BaseCategory = require("../classes/base/BaseCategory.js");
 
 class GuildCategory extends BaseCategory {
 	static code = "guild";
