+++ lib/classes/CommandsRegistry.js
@@ -1,8 +1,8 @@
 const fs = require("fs").promises;
 const { fileExtension, createModuleLog } = require("./Utils.js");
 const log = createModuleLog("Commands");
-const BaseCommand = require("../commands/BaseCommand.js");
-const BaseCategory = require("../categories/BaseCategory.js");
+const BaseCommand = require("./base/BaseCommand.js");
+const BaseCategory = require("./base/BaseCategory.js");
 
 /**
  * # Command Registry
