+++ lib/classes/MessageParser.js
@@ -1,6 +1,6 @@
 const CommandsRegistry = require("./CommandsRegistry.js");
-const BaseCommand = require("../commands/BaseCommand.js");
-const BaseError = require("../classes/errors/BaseError.js");
+const BaseCommand = require("./base/BaseCommand.js");
+const BaseError = require("./base/BaseError.js");
 const CommandContext = require("./CommandContext.js");
 const Lang = require("./Lang.js");
 const { Guild, GuildChannel, GuildMember, User, MessageStatistics } = require("./Database.js");
