+++ lib/classes/base/BaseCategory.js
