+++ lib/classes/base/BaseCommand.js
@@ -1,5 +1,5 @@
-const CommandContextValidator = require("../classes/validation/CommandContextValidator.js");
-const CommandContext = require("../classes/CommandContext.js");
+const CommandContextValidator = require("../validation/CommandContextValidator.js");
+const CommandContext = require("../CommandContext.js");
 
 /**
  * # Base Command
