+++ lib/classes/base/BaseError.js
