+++ lib/classes/errors/ArgumentError.js
@@ -1,4 +1,4 @@
-const BaseError = require("./BaseError.js");
+const BaseError = require("../base/BaseError.js");
 
 class ArgumentError extends BaseError {
 	static errorCode = "invalidArgumentError";
