+++ lib/classes/errors/CallerPermissionError.js
@@ -1,4 +1,4 @@
-const BaseError = require("./BaseError.js");
+const BaseError = require("../base/BaseError.js");
 
 class CallerPermissionError extends BaseError {
 	static errorCode = "callerPermissionError";
