+++ lib/commands/basic/AboutCommand.js
@@ -1,4 +1,4 @@
-const BaseCommand = require("../BaseCommand");
+const BaseCommand = require("../../classes/base/BaseCommand");
 const DefaultEmbed = require("../../classes/embed/DefaultEmbed");
 const CoreCategory = require("../../categories/CoreCategory");
