+++ lib/commands/basic/HelpCommand.js
@@ -1,4 +1,4 @@
-const BaseCommand = require("../BaseCommand.js");
+const BaseCommand = require("../../classes/base/BaseCommand.js");
 const CommandsRegistry = require("../../classes/CommandsRegistry.js");
 const CoreCategory = require("../../categories/CoreCategory.js");
 const DefaultEmbed = require("../../classes/embed/DefaultEmbed.js");
