+++ lib/commands/basic/PingCommand.js
@@ -1,5 +1,5 @@
 const Bot = require("../../classes/Bot.js");
-const BaseCommand = require("../BaseCommand.js");
+const BaseCommand = require("../../classes/base/BaseCommand.js");
 const CoreCategory = require("../../categories/CoreCategory.js");
 const DefaultEmbed = require("../../classes/embed/DefaultEmbed.js");
