+++ lib/commands/guild/PrefixCommand.js
@@ -1,4 +1,4 @@
-const BaseCommand = require("../BaseCommand.js");
+const BaseCommand = require("../../classes/base/BaseCommand.js");
 const DefaultEmbed = require("../../classes/embed/DefaultEmbed.js");
 const ConfigManager = require("../../classes/ConfigManager.js");
 const GuildCategory = require("../../categories/GuildCategory.js");
