+++ lib/commands/guild/VoiceMergeCommand.js
@@ -1,4 +1,4 @@
-const BaseCommand = require("../BaseCommand.js");
+const BaseCommand = require("../../classes/base/BaseCommand.js");
 const GuildCategory = require("../../categories/GuildCategory.js");
 
 class VoiceMergeCommand extends BaseCommand {
