+++ lib/commands/guild/WelcomeCommand.js
@@ -1,4 +1,4 @@
-const BaseCommand = require("../BaseCommand.js");
+const BaseCommand = require("../../classes/base/BaseCommand.js");
 const GuildCategory = require("../../categories/GuildCategory.js");
 const DefaultEmbed = require("../../classes/embed/DefaultEmbed.js");
